Rework Command Invocation
